Split metadata and vocabulary providers
