+++ projects/mercury/src/components/metadata/LinkedDataWrapper.js
@@ -1,7 +1,8 @@
 import React from 'react';
 
 import {METADATA_PATH, VOCABULARY_PATH} from "../../constants";
-import {LinkedDataVocabularyProviderContainer, LinkedDataMetadataProviderContainer} from './LinkedDataContext';
+import LinkedDataVocabularyProvider from './LinkedDataVocabularyProvider';
+import LinkedDataMetadataProvider from './LinkedDataMetadataProvider';
 import {LinkedDataValuesContext} from "./common/LinkedDataValuesContext";
 import MetadataValueComponentFactory from "./metadata/MetadataValueComponentFactory";
 import VocabularyValueComponentFactory from "./vocabulary/VocabularyValueComponentFactory";
@@ -9,20 +10,20 @@ import BreadcrumbsContext from '../common/breadcrumbs/BreadcrumbsContext';
 
 export const MetadataWrapper = ({children}) => (
     <BreadcrumbsContext.Provider value={{segments: [{label: 'Metadata', href: '/metadata', icon: 'assignment'}]}}>
-        <LinkedDataMetadataProviderContainer>
+        <LinkedDataMetadataProvider>
             <LinkedDataValuesContext.Provider value={{editorPath: METADATA_PATH, componentFactory: MetadataValueComponentFactory}}>
                 {children}
             </LinkedDataValuesContext.Provider>
-        </LinkedDataMetadataProviderContainer>
+        </LinkedDataMetadataProvider>
     </BreadcrumbsContext.Provider>
 );
 
 export const VocabularyWrapper = ({children}) => (
     <BreadcrumbsContext.Provider value={{segments: [{label: 'Vocabulary', href: '/vocabulary', icon: 'code'}]}}>
-        <LinkedDataVocabularyProviderContainer>
+        <LinkedDataVocabularyProvider>
             <LinkedDataValuesContext.Provider value={{editorPath: VOCABULARY_PATH, componentFactory: VocabularyValueComponentFactory}}>
                 {children}
             </LinkedDataValuesContext.Provider>
-        </LinkedDataVocabularyProviderContainer>
+        </LinkedDataVocabularyProvider>
     </BreadcrumbsContext.Provider>
 );
